@@ -16,8 +16,10 @@
 
 package com.example.android.architecture.blueprints.todoapp.taskdetail;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.example.android.architecture.blueprints.todoapp.UseCase;
 import com.example.android.architecture.blueprints.todoapp.UseCaseHandler;
@@ -27,8 +29,6 @@ import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Ta
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase.ActivateTask;
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase.CompleteTask;
 import com.google.common.base.Strings;
-
-import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Listens to user actions from the UI ({@link TaskDetailFragment}), retrieves the data and updates
@@ -46,13 +46,7 @@ public class TaskDetailPresenter implements TaskDetailContract.Presenter {
     @Nullable
     private String mTaskId;
 
-    public TaskDetailPresenter(@NonNull UseCaseHandler useCaseHandler,
-            @Nullable String taskId,
-            @NonNull TaskDetailContract.View taskDetailView,
-            @NonNull GetTask getTask,
-            @NonNull CompleteTask completeTask,
-            @NonNull ActivateTask activateTask,
-            @NonNull DeleteTask deleteTask) {
+    public TaskDetailPresenter(@NonNull UseCaseHandler useCaseHandler, @Nullable String taskId, @NonNull TaskDetailContract.View taskDetailView, @NonNull GetTask getTask, @NonNull CompleteTask completeTask, @NonNull ActivateTask activateTask, @NonNull DeleteTask deleteTask) {
         mTaskId = taskId;
         mUseCaseHandler = checkNotNull(useCaseHandler, "useCaseHandler cannot be null!");
         mTaskDetailView = checkNotNull(taskDetailView, "taskDetailView cannot be null!");
@@ -76,29 +70,28 @@ public class TaskDetailPresenter implements TaskDetailContract.Presenter {
 
         mTaskDetailView.setLoadingIndicator(true);
 
-        mUseCaseHandler.execute(mGetTask, new GetTask.RequestValues(mTaskId),
-                new UseCase.UseCaseCallback<GetTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(GetTask.ResponseValue response) {
-                        Task task = response.getTask();
+        mUseCaseHandler.execute(mGetTask, new GetTask.RequestValues(mTaskId), new UseCase.UseCaseCallback<GetTask.ResponseValue>() {
+            @Override
+            public void onSuccess(GetTask.ResponseValue response) {
+                Task task = response.getTask();
 
-                        // The view may not be able to handle UI updates anymore
-                        if (!mTaskDetailView.isActive()) {
-                            return;
-                        }
-                        mTaskDetailView.setLoadingIndicator(false);
-                        showTask(task);
-                    }
+                // The view may not be able to handle UI updates anymore
+                if (!mTaskDetailView.isActive()) {
+                    return;
+                }
+                mTaskDetailView.setLoadingIndicator(false);
+                showTask(task);
+            }
 
-                    @Override
-                    public void onError() {
-                        // The view may not be able to handle UI updates anymore
-                        if (!mTaskDetailView.isActive()) {
-                            return;
-                        }
-                        mTaskDetailView.showMissingTask();
-                    }
-                });
+            @Override
+            public void onError() {
+                // The view may not be able to handle UI updates anymore
+                if (!mTaskDetailView.isActive()) {
+                    return;
+                }
+                mTaskDetailView.showMissingTask();
+            }
+        });
     }
 
     @Override
@@ -112,18 +105,17 @@ public class TaskDetailPresenter implements TaskDetailContract.Presenter {
 
     @Override
     public void deleteTask() {
-        mUseCaseHandler.execute(mDeleteTask, new DeleteTask.RequestValues(mTaskId),
-                new UseCase.UseCaseCallback<DeleteTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(DeleteTask.ResponseValue response) {
-                        mTaskDetailView.showTaskDeleted();
-                    }
+        mUseCaseHandler.execute(mDeleteTask, new DeleteTask.RequestValues(mTaskId), new UseCase.UseCaseCallback<DeleteTask.ResponseValue>() {
+            @Override
+            public void onSuccess(DeleteTask.ResponseValue response) {
+                mTaskDetailView.showTaskDeleted();
+            }
 
-                    @Override
-                    public void onError() {
-                        // Show error, log, etc.
-                    }
-                });
+            @Override
+            public void onError() {
+                // Show error, log, etc.
+            }
+        });
     }
 
     @Override
@@ -133,18 +125,17 @@ public class TaskDetailPresenter implements TaskDetailContract.Presenter {
             return;
         }
 
-        mUseCaseHandler.execute(mCompleteTask, new CompleteTask.RequestValues(mTaskId),
-                new UseCase.UseCaseCallback<CompleteTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(CompleteTask.ResponseValue response) {
-                        mTaskDetailView.showTaskMarkedComplete();
-                    }
+        mUseCaseHandler.execute(mCompleteTask, new CompleteTask.RequestValues(mTaskId), new UseCase.UseCaseCallback<CompleteTask.ResponseValue>() {
+            @Override
+            public void onSuccess(CompleteTask.ResponseValue response) {
+                mTaskDetailView.showTaskMarkedComplete();
+            }
 
-                    @Override
-                    public void onError() {
-                        // Show error, log, etc.
-                    }
-                });
+            @Override
+            public void onError() {
+                // Show error, log, etc.
+            }
+        });
     }
 
     @Override
@@ -153,18 +144,17 @@ public class TaskDetailPresenter implements TaskDetailContract.Presenter {
             mTaskDetailView.showMissingTask();
             return;
         }
-        mUseCaseHandler.execute(mActivateTask, new ActivateTask.RequestValues(mTaskId),
-                new UseCase.UseCaseCallback<ActivateTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(ActivateTask.ResponseValue response) {
-                        mTaskDetailView.showTaskMarkedActive();
-                    }
+        mUseCaseHandler.execute(mActivateTask, new ActivateTask.RequestValues(mTaskId), new UseCase.UseCaseCallback<ActivateTask.ResponseValue>() {
+            @Override
+            public void onSuccess(ActivateTask.ResponseValue response) {
+                mTaskDetailView.showTaskMarkedActive();
+            }
 
-                    @Override
-                    public void onError() {
-                        // Show error, log, etc.
-                    }
-                });
+            @Override
+            public void onError() {
+                // Show error, log, etc.
+            }
+        });
     }
 
     private void showTask(@NonNull Task task) {
