@@ -16,7 +16,8 @@
 
 package com.example.android.architecture.blueprints.todoapp.tasks;
 
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.example.android.architecture.blueprints.todoapp.BaseView;
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
