@@ -19,13 +19,14 @@ package com.example.android.architecture.blueprints.todoapp.tasks;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import android.app.Activity;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.example.android.architecture.blueprints.todoapp.UseCase;
 import com.example.android.architecture.blueprints.todoapp.UseCaseHandler;
 import com.example.android.architecture.blueprints.todoapp.addedittask.AddEditTaskActivity;
-import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource;
+import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase.ActivateTask;
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase.ClearCompleteTasks;
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase.CompleteTask;
@@ -52,17 +53,13 @@ public class TasksPresenter implements TasksContract.Presenter {
 
     private final UseCaseHandler mUseCaseHandler;
 
-    public TasksPresenter(@NonNull UseCaseHandler useCaseHandler,
-            @NonNull TasksContract.View tasksView, @NonNull GetTasks getTasks,
-            @NonNull CompleteTask completeTask, @NonNull ActivateTask activateTask,
-            @NonNull ClearCompleteTasks clearCompleteTasks) {
+    public TasksPresenter(@NonNull UseCaseHandler useCaseHandler, @NonNull TasksContract.View tasksView, @NonNull GetTasks getTasks, @NonNull CompleteTask completeTask, @NonNull ActivateTask activateTask, @NonNull ClearCompleteTasks clearCompleteTasks) {
         mUseCaseHandler = checkNotNull(useCaseHandler, "usecaseHandler cannot be null");
         mTasksView = checkNotNull(tasksView, "tasksView cannot be null!");
         mGetTasks = checkNotNull(getTasks, "getTask cannot be null!");
         mCompleteTask = checkNotNull(completeTask, "completeTask cannot be null!");
         mActivateTask = checkNotNull(activateTask, "activateTask cannot be null!");
-        mClearCompleteTasks = checkNotNull(clearCompleteTasks,
-                "clearCompleteTasks cannot be null!");
+        mClearCompleteTasks = checkNotNull(clearCompleteTasks, "clearCompleteTasks cannot be null!");
 
 
         mTasksView.setPresenter(this);
@@ -76,8 +73,7 @@ public class TasksPresenter implements TasksContract.Presenter {
     @Override
     public void result(int requestCode, int resultCode) {
         // If a task was successfully added, show snackbar
-        if (AddEditTaskActivity.REQUEST_ADD_TASK == requestCode
-                && Activity.RESULT_OK == resultCode) {
+        if (AddEditTaskActivity.REQUEST_ADD_TASK == requestCode && Activity.RESULT_OK == resultCode) {
             mTasksView.showSuccessfullySavedMessage();
         }
     }
@@ -98,34 +94,32 @@ public class TasksPresenter implements TasksContract.Presenter {
             mTasksView.setLoadingIndicator(true);
         }
 
-        GetTasks.RequestValues requestValue = new GetTasks.RequestValues(forceUpdate,
-                mCurrentFiltering);
+        GetTasks.RequestValues requestValue = new GetTasks.RequestValues(forceUpdate, mCurrentFiltering);
 
-        mUseCaseHandler.execute(mGetTasks, requestValue,
-                new UseCase.UseCaseCallback<GetTasks.ResponseValue>() {
-                    @Override
-                    public void onSuccess(GetTasks.ResponseValue response) {
-                        List<Task> tasks = response.getTasks();
-                        // The view may not be able to handle UI updates anymore
-                        if (!mTasksView.isActive()) {
-                            return;
-                        }
-                        if (showLoadingUI) {
-                            mTasksView.setLoadingIndicator(false);
-                        }
+        mUseCaseHandler.execute(mGetTasks, requestValue, new UseCase.UseCaseCallback<GetTasks.ResponseValue>() {
+            @Override
+            public void onSuccess(GetTasks.ResponseValue response) {
+                List<Task> tasks = response.getTasks();
+                // The view may not be able to handle UI updates anymore
+                if (!mTasksView.isActive()) {
+                    return;
+                }
+                if (showLoadingUI) {
+                    mTasksView.setLoadingIndicator(false);
+                }
 
-                        processTasks(tasks);
-                    }
+                processTasks(tasks);
+            }
 
-                    @Override
-                    public void onError() {
-                        // The view may not be able to handle UI updates anymore
-                        if (!mTasksView.isActive()) {
-                            return;
-                        }
-                        mTasksView.showLoadingTasksError();
-                    }
-                });
+            @Override
+            public void onError() {
+                // The view may not be able to handle UI updates anymore
+                if (!mTasksView.isActive()) {
+                    return;
+                }
+                mTasksView.showLoadingTasksError();
+            }
+        });
     }
 
     private void processTasks(List<Task> tasks) {
@@ -182,55 +176,51 @@ public class TasksPresenter implements TasksContract.Presenter {
     @Override
     public void completeTask(@NonNull Task completedTask) {
         checkNotNull(completedTask, "completedTask cannot be null!");
-        mUseCaseHandler.execute(mCompleteTask, new CompleteTask.RequestValues(
-                        completedTask.getId()),
-                new UseCase.UseCaseCallback<CompleteTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(CompleteTask.ResponseValue response) {
-                        mTasksView.showTaskMarkedComplete();
-                        loadTasks(false, false);
-                    }
+        mUseCaseHandler.execute(mCompleteTask, new CompleteTask.RequestValues(completedTask.getId()), new UseCase.UseCaseCallback<CompleteTask.ResponseValue>() {
+            @Override
+            public void onSuccess(CompleteTask.ResponseValue response) {
+                mTasksView.showTaskMarkedComplete();
+                loadTasks(false, false);
+            }
 
-                    @Override
-                    public void onError() {
-                        mTasksView.showLoadingTasksError();
-                    }
-                });
+            @Override
+            public void onError() {
+                mTasksView.showLoadingTasksError();
+            }
+        });
     }
 
     @Override
     public void activateTask(@NonNull Task activeTask) {
         checkNotNull(activeTask, "activeTask cannot be null!");
-        mUseCaseHandler.execute(mActivateTask, new ActivateTask.RequestValues(activeTask.getId()),
-                new UseCase.UseCaseCallback<ActivateTask.ResponseValue>() {
-                    @Override
-                    public void onSuccess(ActivateTask.ResponseValue response) {
-                        mTasksView.showTaskMarkedActive();
-                        loadTasks(false, false);
-                    }
+        mUseCaseHandler.execute(mActivateTask, new ActivateTask.RequestValues(activeTask.getId()), new UseCase.UseCaseCallback<ActivateTask.ResponseValue>() {
+            @Override
+            public void onSuccess(ActivateTask.ResponseValue response) {
+                mTasksView.showTaskMarkedActive();
+                loadTasks(false, false);
+            }
 
-                    @Override
-                    public void onError() {
-                        mTasksView.showLoadingTasksError();
-                    }
-                });
+            @Override
+            public void onError() {
+                mTasksView.showLoadingTasksError();
+            }
+        });
     }
 
     @Override
     public void clearCompletedTasks() {
-        mUseCaseHandler.execute(mClearCompleteTasks, new ClearCompleteTasks.RequestValues(),
-                new UseCase.UseCaseCallback<ClearCompleteTasks.ResponseValue>() {
-                    @Override
-                    public void onSuccess(ClearCompleteTasks.ResponseValue response) {
-                        mTasksView.showCompletedTasksCleared();
-                        loadTasks(false, false);
-                    }
+        mUseCaseHandler.execute(mClearCompleteTasks, new ClearCompleteTasks.RequestValues(), new UseCase.UseCaseCallback<ClearCompleteTasks.ResponseValue>() {
+            @Override
+            public void onSuccess(ClearCompleteTasks.ResponseValue response) {
+                mTasksView.showCompletedTasksCleared();
+                loadTasks(false, false);
+            }
 
-                    @Override
-                    public void onError() {
-                        mTasksView.showLoadingTasksError();
-                    }
-                });
+            @Override
+            public void onError() {
+                mTasksView.showLoadingTasksError();
+            }
+        });
     }
 
     /**
