@@ -16,12 +16,12 @@
 
 package com.example.android.architecture.blueprints.todoapp.tasks.domain.model;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Ignore;
-import android.arch.persistence.room.PrimaryKey;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.Ignore;
+import androidx.room.PrimaryKey;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Strings;
@@ -95,8 +95,7 @@ public final class Task {
      * @param id          id of the task
      * @param completed   true if the task is completed, false if it's active
      */
-    public Task(@Nullable String title, @Nullable String description,
-                @NonNull String id, boolean completed) {
+    public Task(@Nullable String title, @Nullable String description, @NonNull String id, boolean completed) {
         mId = id;
         mTitle = title;
         mDescription = description;
@@ -136,8 +135,7 @@ public final class Task {
     }
 
     public boolean isEmpty() {
-        return Strings.isNullOrEmpty(mTitle) &&
-               Strings.isNullOrEmpty(mDescription);
+        return Strings.isNullOrEmpty(mTitle) && Strings.isNullOrEmpty(mDescription);
     }
 
     @Override
@@ -145,9 +143,7 @@ public final class Task {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Task task = (Task) o;
-        return Objects.equal(mId, task.mId) &&
-               Objects.equal(mTitle, task.mTitle) &&
-               Objects.equal(mDescription, task.mDescription);
+        return Objects.equal(mId, task.mId) && Objects.equal(mTitle, task.mTitle) && Objects.equal(mDescription, task.mDescription);
     }
 
     @Override
