@@ -16,12 +16,12 @@
 
 package com.example.android.architecture.blueprints.todoapp.tasks.domain.usecase;
 
-import android.support.annotation.NonNull;
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import androidx.annotation.NonNull;
 
 import com.example.android.architecture.blueprints.todoapp.UseCase;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository;
-
-import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Marks a task as completed.
