@@ -16,10 +16,11 @@
 
 package com.example.android.architecture.blueprints.todoapp.data.source.local;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
+
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 
 import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 
@@ -27,7 +28,7 @@ import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Ta
 /**
  * The Room Database that contains the Task table.
  */
-@Database(entities = {Task.class}, version = 1)
+@Database(entities = {Task.class}, version = 1, exportSchema = false)
 public abstract class ToDoDatabase extends RoomDatabase {
 
     private static ToDoDatabase INSTANCE;
@@ -39,9 +40,7 @@ public abstract class ToDoDatabase extends RoomDatabase {
     public static ToDoDatabase getInstance(Context context) {
         synchronized (sLock) {
             if (INSTANCE == null) {
-                INSTANCE = Room.databaseBuilder(context.getApplicationContext(),
-                        ToDoDatabase.class, "Tasks.db")
-                        .build();
+                INSTANCE = Room.databaseBuilder(context.getApplicationContext(), ToDoDatabase.class, "Tasks.db").build();
             }
             return INSTANCE;
         }
